VNU-22 address review submission feedback
