--- src/lib/types/reviews.ts
@@ -1,15 +1,0 @@
-export type CreateProductReviewInput = {
-  rating: number
-  title?: string
-  content?: string
-  reviewer_name?: string
-  reviewer_email?: string
-  metadata?: Record<string, unknown>
-}
-
-export type ProductReviewSubmissionResult = {
-  productId: string
-  alreadyReviewed?: boolean
-  message?: string
-  success?: boolean
-}
