+++ src/routes/ordrer/[id]/anmeld/+page.server.ts
@@ -1,58 +1,16 @@
-import type { Actions, PageServerLoad } from './$types'
+import type { Actions } from './$types'
 import { error, fail } from '@sveltejs/kit'
 import type { HttpTypes } from '@medusajs/types'
 import { sdk } from '$lib/medusa'
-import type { CreateProductReviewInput } from '$lib/types/reviews'
+import type { CreateProductReviewInput } from '../../../api/products/[id]/reviews/+server'
+import { getReviewableItems, type ReviewableOrderItem } from './reviewable-items'
 
 const MAX_TITLE_LENGTH = 120
 const MAX_CONTENT_LENGTH = 5000
 const MAX_REVIEWER_NAME_LENGTH = 120
 
-type ReviewableOrderItem = {
-  lineItemIds: string[]
-  productHandle: string | null
-  productId: string
-  quantity: number
-  thumbnail: string | null
-  title: string
-  variantTitle: string | null
-}
-
 function getString(value: FormDataEntryValue | null) {
   return typeof value === 'string' ? value.trim() : ''
-}
-
-function getReviewerName(order: HttpTypes.StoreOrder) {
-  const address = order.shipping_address ?? order.billing_address
-  return [address?.first_name, address?.last_name].filter(Boolean).join(' ')
-}
-
-function getReviewableItems(order: HttpTypes.StoreOrder): ReviewableOrderItem[] {
-  const itemsByProductId = new Map<string, ReviewableOrderItem>()
-
-  for (const item of order.items ?? []) {
-    if (!item.product_id) continue
-
-    const existingItem = itemsByProductId.get(item.product_id)
-
-    if (existingItem) {
-      existingItem.lineItemIds.push(item.id)
-      existingItem.quantity += item.quantity
-      continue
-    }
-
-    itemsByProductId.set(item.product_id, {
-      lineItemIds: [item.id],
-      productHandle: item.product_handle,
-      productId: item.product_id,
-      quantity: item.quantity,
-      thumbnail: item.thumbnail,
-      title: item.product_title ?? item.title,
-      variantTitle: item.variant_title,
-    })
-  }
-
-  return Array.from(itemsByProductId.values())
 }
 
 async function getOrder(orderId: string) {
@@ -129,22 +87,6 @@ function getReviewPayload(
       reviewer_name: reviewerName,
       title,
     },
-  }
-}
-
-export const load: PageServerLoad = async ({ locals, params }) => {
-  const order = await getOrder(params.id)
-  const reviewableItems = getReviewableItems(order)
-
-  return {
-    locale: locals.locale,
-    order: {
-      display_id: order.display_id,
-      email: order.email,
-      id: order.id,
-    },
-    reviewerName: getReviewerName(order),
-    reviewableItems,
   }
 }
